+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - [#6713](https://github.com/influxdata/influxdb/pull/6713): Reduce allocations during query parsing.
 - [#3733](https://github.com/influxdata/influxdb/issues/3733): Modify the default retention policy name and make it configurable.
 - [#5655](https://github.com/influxdata/influxdb/issues/5655): Support specifying a retention policy for the graphite service.
+- [#6820](https://github.com/influxdata/influxdb/issues/6820): Add NodeID to execution options
 
 ### Bugfixes